requirements.txt
flask==2.0.1
flask-cors==3.0.10
pdf2docx==0.5.6
Pillow==9.0.0
pytesseract==0.3.8
pandas==1.3.3
python-docx==0.8.11
pdf2image==1.16.0
gunicorn==20.1.0